Combine DeleteNotificationsButton and ReadNotificationsButton components into one
To duplicate SonarCloud duplication issues
